Find node details by name
If a node contains and invalid `provider-id` label we fallback to
trying to find the node details by name. This is not a good idea as it
can hit API rate limits on larger clusters. A better solution if to
use another label to try and find the node ID.